Move api to pkg/sdk to trim dependencies
