+++ go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/banzaicloud/istio-client-go v0.0.0-20200410173743-e1adddc949b3
 	github.com/banzaicloud/istio-operator v0.0.0-20200407070503-3f7dc6953a7b
 	github.com/banzaicloud/k8s-objectmatcher v1.1.1
+	github.com/banzaicloud/kafka-operator/api v0.0.0
 	github.com/envoyproxy/go-control-plane v0.9.4
 	github.com/ghodss/yaml v1.0.1-0.20190212211648-25d852aebe32
 	github.com/go-logr/logr v0.1.0
@@ -20,7 +21,7 @@ require (
 	github.com/hashicorp/vault/sdk v0.1.14-0.20200406173424-43a93d4a43b1
 	github.com/imdario/mergo v0.3.9
 	github.com/influxdata/influxdb v1.7.6 // indirect
-	github.com/jetstack/cert-manager v0.14.1
+	github.com/jetstack/cert-manager v0.14.2
 	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
 	github.com/lestrrat-go/backoff v1.0.0
 	github.com/lyft/protoc-gen-validate v0.1.0 // indirect
@@ -35,3 +36,5 @@ require (
 	k8s.io/client-go v0.17.3
 	sigs.k8s.io/controller-runtime v0.5.2
 )
+
+replace github.com/banzaicloud/kafka-operator/api => ./pkg/sdk
