+++ pkg/sdk/go.mod
@@ -1,0 +1,11 @@
+module github.com/banzaicloud/kafka-operator/pkg/sdk
+
+go 1.13
+
+require (
+	github.com/banzaicloud/istio-client-go v0.0.0-20200410173743-e1adddc949b3
+	github.com/jetstack/cert-manager v0.14.2
+	k8s.io/api v0.17.3
+	k8s.io/apimachinery v0.17.3
+	sigs.k8s.io/controller-runtime v0.5.2
+)
