+++ pkg/sdk/v1alpha1/kafkauser_types.go
@@ -18,7 +18,7 @@ import (
 	cmmeta "github.com/jetstack/cert-manager/pkg/apis/meta/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/banzaicloud/kafka-operator/api/v1beta1"
+	"github.com/banzaicloud/kafka-operator/pkg/sdk/v1beta1"
 )
 
 // KafkaUserSpec defines the desired state of KafkaUser
